+++ tests/test_entities.py
@@ -70,7 +70,6 @@ async def test_account_sensors(
         DOMAIN,
         f"{entry.entry_id}_account_raw_money",
     )
-    raw_money_entity = entity_registry.async_get(raw_money_entity_id)
     green_stacks_entity_id = entity_registry.async_get_entity_id(
         "sensor",
         DOMAIN,
@@ -120,11 +119,6 @@ async def test_account_sensors(
         "sensor",
         DOMAIN,
         f"{entry.entry_id}_account_pets",
-    )
-    achievements_by_world_entity_id = entity_registry.async_get_entity_id(
-        "sensor",
-        DOMAIN,
-        f"{entry.entry_id}_account_achievements_by_world",
     )
     task_levels_entity_id = entity_registry.async_get_entity_id(
         "sensor",
@@ -189,7 +183,7 @@ async def test_account_sensors(
     killroy_entity_id = entity_registry.async_get_entity_id(
         "sensor",
         DOMAIN,
-        f"{entry.entry_id}_account_killroy",
+        f"{entry.entry_id}_account_world_2_killroy",
     )
 
     assert hass.states.get(total_level_entity_id).state == "365"
@@ -202,8 +196,7 @@ async def test_account_sensors(
     assert hass.states.get(highest_level_entity_id).state == "210"
     assert hass.states.get(total_skill_entity_id).state == "205"
     assert hass.states.get(total_money_entity_id).state == "987654"
-    assert raw_money_entity.disabled_by is er.RegistryEntryDisabler.INTEGRATION
-    assert hass.states.get(raw_money_entity_id) is None
+    assert hass.states.get(raw_money_entity_id).state == "987654"
     assert hass.states.get(green_stacks_entity_id).state == "3"
     assert hass.states.get(slab_entity_id).state == "456"
     assert hass.states.get(achievements_entity_id).state == "78"
@@ -214,7 +207,6 @@ async def test_account_sensors(
     assert hass.states.get(minigame_scores_entity_id).state == "1827"
     assert hass.states.get(progress_totals_entity_id).state == "11"
     assert hass.states.get(pets_entity_id).state == "3"
-    assert hass.states.get(achievements_by_world_entity_id).state == "78"
     assert hass.states.get(task_levels_entity_id).state == "2"
     assert hass.states.get(taskboard_merits_entity_id).state == "2"
     assert hass.states.get(taskboard_unlocks_entity_id).state == "3"
@@ -227,7 +219,7 @@ async def test_account_sensors(
     assert hass.states.get(world_2_bubbles_entity_id).state == "1"
     assert hass.states.get(world_2_sigils_entity_id).state == "1"
     assert hass.states.get(world_2_vote_ballots_entity_id).state == "1"
-    assert hass.states.get(killroy_entity_id).state == "2"
+    assert hass.states.get(killroy_entity_id).state == "3"
     highest_level_attributes = hass.states.get(highest_level_entity_id).attributes
     assert highest_level_attributes["highest_level_character"] == "Bubo Main"
     assert highest_level_attributes["class_counts"] == {
@@ -287,9 +279,9 @@ async def test_account_sensors(
         "Slime": "0/1",
     }
     assert (
-        hass.states.get(achievements_by_world_entity_id).attributes[
-            "achievement_status"
-        ]["World 1"]["progress"]["Achievement Hunter"]
+        hass.states.get(achievements_entity_id).attributes["achievement_status"][
+            "World 1"
+        ]["progress"]["Achievement Hunter"]
         == 66
     )
     assert (
@@ -362,7 +354,10 @@ async def test_account_sensors(
         == 1
     )
     assert (
-        hass.states.get(killroy_entity_id).attributes["killroy"]["rooms_available"] == 3
+        hass.states.get(killroy_entity_id).attributes["world_2_killroy"][
+            "rooms_available"
+        ]
+        == 3
     )
     assert last_updated_entity.entity_category is EntityCategory.DIAGNOSTIC
     assert last_updated_entity.disabled_by is er.RegistryEntryDisabler.INTEGRATION